backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
gunicorn==21.2.0
pytest==8.3.3
SQLAlchemy==2.0.44
python-dotenv==1.2.1
